minor fixes to live server

+++ server/db/migrations/20231223184722_create_pack_items.ts
@@ -23,7 +23,7 @@ export async function up(knex: Knex): Promise<void> {
     table.string("pack_item_name").notNullable();
     table.text("pack_item_description").nullable();
     table.integer("pack_item_quantity").defaultTo(1).notNullable();
-    table.bigint("pack_item_weight").nullable().defaultTo(0);
+    table.decimal("pack_item_weight").nullable().defaultTo(0);
     table.string("pack_item_unit").notNullable().defaultTo("oz");
     table.bigint("pack_item_price").nullable().defaultTo(0);
     table.string("pack_item_url").nullable();
